+++ js/data.js
@@ -1,3 +1,5 @@
+import { getRandomNumber} from './util.js';
+
 const description = ['Я отстаю!!!','Скоро лето, скоро вэйк','горы и сноуборд','прыгаю с парашютом','Работаю работу','учу js','мой котЭ туповат'];
 const userName = ['Абрам','Аваз','Август','Авдей','Автандил','Адам','Адис','Адольф','Адриан'];
 const message = ['Всё отлично!',
@@ -7,4 +9,27 @@ const message = ['Всё отлично!',
   'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
   'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
 
-export {description, userName, message};
+const getRandomArrayElement = (element) => element[getRandomNumber(0, element.length - 1)];
+
+const avatar = Array.from({length: 6}, (_, i) => `img/avatar-${i+1}.svg`);
+
+const comment = (userId) => ({
+  id : userId,
+  avatar : getRandomArrayElement(avatar),
+  name : getRandomArrayElement(userName),
+  message : getRandomArrayElement(message)
+});
+
+const url = Array.from({length: 25}, (_, i) => `photos/${i+1}.jpg`);
+
+const createObject = (id) => ({
+  id : id,
+  url : getRandomArrayElement(url),
+  description : getRandomArrayElement(description),
+  likes : getRandomNumber(15, 200),
+  comments : Array.from({length: getRandomNumber(0,30)}, (_, i) => comment(i+1))
+});
+
+const newArray = Array.from({length: 25}, (_, i) => createObject(i+1));
+
+export {description, userName, message, getRandomArrayElement, newArray, avatar, comment, url, createObject};
